[FEAT] #34 인기 장소 조회 api 연동

+++ app/src/main/java/com/moduro/barrier_free_app/presentation/search/navigation/SearchNavigator.kt
@@ -6,7 +6,7 @@ class SearchNavigator(
     val navController: NavController
 ){
 
-    fun navigateToPlaceDetail(placeId: Int) {
+    fun navigateToPlaceDetail(placeId: Long) {
         navController.navigate("placeDetail/$placeId")
     }
